+++ bin/stats2multiqc.sh
@@ -20,6 +20,7 @@ function help {
     exit;
 }
 
+is_pe=0
 while getopts "s:d:a:ph" OPT
 do
     case $OPT in
@@ -53,54 +54,32 @@ echo -e "Sample_ID,Sample_name,Number_of_reads,Fragment_length,Number_of_aligned
 for sample in $all_samples
 do
     #SAMPLE NAME
-    sname=$(grep "$sample" $splan | awk -F, '{print $2}')
+    sname=$(grep "$sample," $splan | awk -F, '{print $2}')
 
     #ALIGNMENT
     if [ $aligner == "bowtie2" ]; then
 	nb_reads=$(grep "reads;" mapping/${sample}_bowtie2.log | sed 's/ .*//')
-#	if [[ $is_pe == "1" ]]; then
-#           nb_uniq_reads=$(grep "concordantly exactly" mapping/${sample}_bowtie2.log | awk '{print $1}')
-#           perc_uniq_reads=$(echo "${nb_uniq_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
-#           nb_mult_reads=$(grep "concordantly >1" mapping/${sample}_bowtie2.log | awk '{print $1}')
-#           perc_mult_reads=$(echo "${nb_mult_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
-#           nb_mapped=$(($nb_uniq_reads + $nb_mult_reads))
-#           perc_mapped=$(echo "${nb_mapped} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
-#	else 
-#	    nb_uniq_reads=$(grep "exactly" mapping/${sample}_bowtie2.log | awk '{print $1}')
-#	    perc_uniq_reads=$(echo "${nb_uniq_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
-#	    nb_mult_reads=$(grep ">1" mapping/${sample}_bowtie2.log | awk '{print $1}')
-#	    perc_mult_reads=$(echo "${nb_mult_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
-#	    nb_mapped=$(($nb_uniq_reads + $nb_mult_reads))
-#	    perc_mapped=$(echo "${nb_mapped} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
-#	fi
     elif [ $aligner == "bwa-mem" ]; then
 	nb_reads=$(grep 'Total' mapping/${sample}_bwa.log | awk -F "\t" '{print $2}')
 	tail -n +3 mapping/${sample}_bwa.log > mapping/${sample}_bwa.mqc
-#	nb_unmapped=$(grep 'Unmapped' mapping/${sample}_bwa.log | awk -F "\t" '{print $2}')
-#	nb_uniq_reads=$(grep 'Uniquely' mapping/${sample}_bwa.log | awk -F "\t" '{print $2}')
-#	nb_mult_reads=$(grep 'Multi' mapping/${sample}_bwa.log | awk -F "\t" '{print $2}')
-#	nb_reads=$(( $nb_unmapped + $nb_uniq_reads + $nb_mult_reads ))
-#	perc_uniq_reads=$(echo "${nb_uniq_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
-#	perc_mult_reads=$(echo "${nb_mult_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
-#	nb_mapped=$(($nb_uniq_reads + $nb_mult_reads))
-#	perc_mapped=$(echo "${nb_mapped} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
     elif [ $aligner == "star" ]; then
 	nb_reads=$(grep "Number of input reads" mapping/${sample}Log.final.out | cut -d"|" -f 2 | sed -e 's/\t//g')
-#	nb_uniq_reads=$(grep "Uniquely.*number" mapping/${sample}Log.final.out | awk '{print $NF}')
-#	perc_uniq_reads=$(echo "${nb_uniq_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
-#	nb_mult_reads=$(grep "Number.*multiple" mapping/${sample}Log.final.out | awk '{print $NF}')
-#	perc_mult_reads=$(echo "${nb_mult_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
-#	nb_mapped=$(($nb_uniq_reads + $nb_mult_reads))
-#	perc_mapped=$(echo "${nb_mapped} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
     fi
 
-    #Mapping stats
+    #Mapping stats (always in reads - so must be converted for PE)
+    #These statistics are calculated after spike cleaning but before filtering
     nb_mapped=$(awk -F, '$1=="Mapped"{print $2}' mapping/${sample}_mappingstats.mqc)
-    perc_mapped=$(echo "${nb_mapped} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
     nb_mapped_hq=$(awk -F, '$1=="HighQual"{print $2}' mapping/${sample}_mappingstats.mqc)
-    perc_mapped_hq=$(echo "${nb_mapped_hq} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
     nb_mapped_lq=$(awk -F, '$1=="LowQual"{print $2}' mapping/${sample}_mappingstats.mqc)
-    perc_mapped_lq=$(echo "${nb_mapped_lq} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ') 
+
+    if [[ $is_pe == 1 ]]; then
+	nb_mapped=$(( $nb_mapped / 2))
+	nb_mapped_hq=$(( $nb_mapped_hq / 2))
+	nb_mapped_lq=$(( $nb_mapped_lq / 2))
+    fi
+    perc_mapped=$(echo "${nb_mapped} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
+    perc_mapped_hq=$(echo "${nb_mapped_hq} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
+    perc_mapped_lq=$(echo "${nb_mapped_lq} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
 
     #PICARD
     if [[ -e mapping/${sample}.MarkDuplicates.metrics.txt ]]; then
@@ -108,7 +87,7 @@ do
 	nb_dups_single=$(grep -a2 "## METRICS" mapping/${sample}.MarkDuplicates.metrics.txt | tail -1 | awk -F"\t" '{print $6}')
 	nb_dups_optical=$(grep -a2 "## METRICS" mapping/${sample}.MarkDuplicates.metrics.txt | tail -1 | awk -F"\t" '{print $8}')
 	nb_dups=$(($nb_dups_pair + $nb_dups_single + $nb_dups_optical))
-	perc_dups=$(echo "${nb_dups} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
+	perc_dups=$(echo "${nb_dups} ${nb_mapped}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
     else
 	nb_dups='NA'
 	perc_dups='NA'